add basic support for event driven simulation
define EventEngine class to include a time_wheel of event slots.

change Wire class and init_gates function to accomodate for storing a
referece to next upcoming gate in the wire itself. This in return helps
to easily find the next gate to be evaluated when a wire update happens.